+++ src/app/FakeLib/StrongNamingHelper.fs
@@ -13,8 +13,8 @@ type StrongNameParams =
       /// The directory where the Strong naming process will be started.
       WorkingDir : string }
 
-let SN32 = ProgramFilesX86 @@ "Microsoft SDKs/Windows/v8.0A/bin/NETFX 4.0 Tools/sn.exe"
-let SN64 = ProgramFilesX86 @@ "Microsoft SDKs/Windows/v8.0A/bin/NETFX 4.0 Tools/x64/sn.exe"
+let mutable SN32 = ProgramFilesX86 @@ "Microsoft SDKs/Windows/v8.0A/bin/NETFX 4.0 Tools/sn.exe"
+let mutable SN64 = ProgramFilesX86 @@ "Microsoft SDKs/Windows/v8.0A/bin/NETFX 4.0 Tools/x64/sn.exe"
 
 /// Strong naming default parameters
 let StrongNameDefaults = 
@@ -36,3 +36,13 @@ let StrongName setParams command =
     if not ok then failwithf "SN.exe reported errors."
 
     traceEndTask taskName command
+
+/// Registers the given assembly for verification skipping.
+let DisableVerification assembly key =
+    let command = sprintf "-Vr %s,%s" assembly key
+
+    StrongName id command
+
+    // For 64-bit versions of Windows, we also need to run the 64-bit version of the strong-name tool.
+    if Environment.Is64BitOperatingSystem then
+        StrongName (fun p -> { p with ToolPath = SN64 }) command